backend/requirements/base.txt
# Base requirements — shared by all environments.
Django>=6.0,<7.0
djangorestframework>=3.15,<4.0
django-filter>=25.0,<26.0
django-cors-headers>=4.0,<5.0
celery[redis]>=5.4,<6.0
django-celery-beat>=2.7,<3.0
httpx>=0.27,<1.0
gunicorn>=22.0,<24.0
